Extract banners to an external component
